+++ src/main/java/com/example/demo/Dto/orderItem/OrderItemResponseDto.java
@@ -3,12 +3,14 @@ package com.example.demo.Dto.orderItem;
 import com.example.demo.Dto.order.OrderResponseDto;
 import com.example.demo.Dto.product.ProductResponseDto;
 
+import java.math.BigDecimal;
+
 public class OrderItemResponseDto {
 
     public OrderItemResponseDto() {
     }
 
-    public OrderItemResponseDto(ProductResponseDto productResponseDto, OrderResponseDto orderResponseDto, Double unitPrice, Integer quantity) {
+    public OrderItemResponseDto(ProductResponseDto productResponseDto, OrderResponseDto orderResponseDto, BigDecimal unitPrice, Integer quantity) {
         this.productResponseDto = productResponseDto;
         this.orderResponseDto = orderResponseDto;
         this.unitPrice = unitPrice;
@@ -18,7 +20,7 @@ public class OrderItemResponseDto {
     private Long id;
     private ProductResponseDto productResponseDto;
     private OrderResponseDto orderResponseDto;
-    private Double unitPrice;
+    private BigDecimal unitPrice;
     private Integer quantity;
 
     public Long getId() {
@@ -45,11 +47,11 @@ public class OrderItemResponseDto {
         this.orderResponseDto = orderResponseDto;
     }
 
-    public Double getUnitPrice() {
+    public BigDecimal getUnitPrice() {
         return unitPrice;
     }
 
-    public void setUnitPrice(Double unitPrice) {
+    public void setUnitPrice(BigDecimal unitPrice) {
         this.unitPrice = unitPrice;
     }
